feat: verify core binary
